+++ CHANGES.md
@@ -24,6 +24,7 @@
 -   read DateTime fields with millisecond accuracy (#111)
 -   support writing object columns with np.nan values (#118)
 -   add support to write object columns that contain types different than string (#125)
+-   support writing datetime columns (#120)
 
 ### Breaking changes